Update README.md
I'm gonna miss it. This routine of studying python and go, servers and networks, etc. was a really fun one. But if I'm serious about this, I gotta study other things first. When I come back, I'll become a 10x dev.
Until then, this account is as good as dead.